docker/data_generator/requirements.txt
Faker==8.12.1
pandas==1.1.5
kafka-python==2.0.2
